requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.26.post2
Click==7.0
Django==2.0.6
django-celery==3.3.1
django-cors-headers==2.2.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
kombu==3.0.37
MarkupSafe==1.1.1
numpy==1.17.4
pandas==0.25.3
Pillow==7.1.2
psycopg2==2.8.4
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2018.4
redis==3.4.1
six==1.13.0
Werkzeug==0.16.0
xlrd==1.2.0
